Fixed loss of precision with nDisplay overrides

+++ Source/OpenDRIVE/Public/OpenDriveVehicle.h
@@ -28,8 +28,8 @@ private:
 	int PrevRoadId = -1;
 
 protected:
-	float _SpeedOverride = 0.f;
-	float _AccOverride = 0.f;
+	double _SpeedOverride = 0.;
+	double _AccOverride = 0.;
 	double LengthFront() const;
 	double LengthBack() const;
 
@@ -50,8 +50,8 @@ public:
 	FOnNewRoad OnNewRoad;
 
 	// Overrides speed value for cluster slaves, since they don't have PhysX vehicle model
-	void SetSpeedOverride(float Speed) { _SpeedOverride = Speed; }
-	void SetAccOverride(float Acc) { _AccOverride = Acc; }
+	void SetSpeedOverride(double Speed) { _SpeedOverride = Speed; }
+	void SetAccOverride(double Acc) { _AccOverride = Acc; }
 
 	/**
 	* Returns the longitudinal distance (cm, also known as "pk" or "S" distance) to another vehicle
